Move interval exercise to upward interval exercise
